backend/requirements.txt
apturl==0.5.2
asttokens==2.4.1
backcall==0.2.0
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
comm==0.2.2
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
debugpy==1.8.6
decorator==4.4.2
defer==1.0.6
distro==1.4.0
distro-info==0.23+ubuntu1.1
duplicity==0.8.12.0
entrypoints==0.3
executing==2.1.0
fasteners==0.14.1
future==0.18.2
httplib2==0.14.0
idna==2.8
importlib-metadata==8.5.0
ipykernel==6.29.5
ipython==8.12.3
ipython-genutils==0.2.0
jedi==0.19.1
jupyter-client==8.6.3
jupyter-core==5.7.2
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib-inline==0.1.7
monotonic==1.5
mysql-connector-python==9.0.0
nest-asyncio==1.6.0
netifaces==0.10.4
oauthlib==3.1.0
packaging==24.1
paramiko==2.6.0
parso==0.8.4
pexpect==4.6.0
pickleshare==0.7.5
platformdirs==4.3.6
prompt-toolkit==3.0.48
psutil==6.0.0
pure-eval==0.2.3
pycairo==1.16.2
pycups==1.9.73
pygments==2.18.0
PyGObject==3.36.0
PyJWT==1.7.1
PyNaCl==1.3.0
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.9.0.post0
python-debian==0.1.36+ubuntu1.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==26.2.0
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
speedtest-cli==2.1.2
stack-data==0.6.3
systemd-python==234
tornado==6.4.1
traitlets==5.14.3
typing-extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
uv==0.4.18
wadllib==1.3.3
wcwidth==0.2.13
xkit==0.0.0
zipp==3.20.2
